website_station_backend/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
torchvision
python-multipart
pybluez2
requests